refactor(search_suggestor): split heatmap
